requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
motor>=3.5
jinja2>=3.1
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
python-slugify>=8.0
mistune>=3.0
python-frontmatter>=1.1
pydantic-settings>=2.3
aiofiles>=23.0
